% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_helpers.R
\name{createLinCombFormula}
\alias{createLinCombFormula}
\title{Create formulas for linear combinations}
\usage{
createLinCombFormula(vars, coeffvals = NULL, coeffrefs = NULL)
}
\arguments{
\item{vars}{A vector of row & column indices or an excel reference string for the region whose columns contain the variables
corresponding with the coefficients in coeffs}

\item{coeffs}{vector of coefficients for linear combination or references to worksheet cells containing coefficients}
}
\value{
A character vector containing the linear equations
}
\description{
Create formulas for linear combinations of values in columns.
}
\details{
A formula will be created for each row of the area indicated by the vars argument.
Each formula has terms which are products the columns of the area indicated by the vars argument with either
the corresponding coefficient in the coeffvals argument, or if that is not present then the value in the corresponding cell
indicated by the coeffrefs argument. These terms are summed together to produce the final formula.
}
\author{
Ben Veal
}
