% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_template.R
\name{duplicate_template_sheet}
\alias{duplicate_template_sheet}
\title{Duplicate a template within an Excel reporting template workbook}
\usage{
duplicate_template_sheet(wb, template_name, uniq_name)
}
\arguments{
\item{template_name}{path to Excel template}
}
\value{
a list with a reference to the workbook object and vector of
template names.
}
\description{
Modifies the wb object
}
\examples{

\dontrun{
wb_obj <- load_xl_template("templates/report_template.xlsx")
}
}
