% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_helpers.R
\name{getRefsAdjacentToName}
\alias{getRefsAdjacentToName}
\title{Return indices of cell/region adjacent to another region}
\usage{
getRefsAdjacentToName(wb, name, location = "right", shift = c(0, 0),
  size = c(1, 1), errorOnOverlap = TRUE, asAref = FALSE)
}
\arguments{
\item{wb}{A workbook object}

\item{name}{The name of a named region in wb}

\item{location}{A string indicating which side of the named region to place the new region.
One of: "left","right","above"/"up","below"/"down" or just the first letter of one of those words.}

\item{shift}{A pair of integers indicating how much rightward & downward shift to add to the new region (default is c(0,0)).}

\item{size}{A pair of integers indicating the horizontal & vertical length of the new region (default is c(1,1)).}

\item{errorOnOverlap}{Logical indicating whether or not to throw an error if the new region overlaps the named region.}

\item{asAref}{Logical indicating whether to return the region as an excel reference string or a numeric vector.}
}
\value{
A numeric vector or excel reference string of row & column indices for the new region.
}
\description{
Given a named region in an excel workbook this function can be used to calculate the cell indices
for a region located adjacent to the named one.
}
\details{
The return value will be either an excel reference string (as returned by XLConnect::aref), or a vector
of row & column indices (as returned by aref2idx, i.e. first 2 elements are row & column indices of top-left corner
and last 2 elements are row & column indices of bottom-right corner).
By default a reference to a single cell will be returned, but if size is set to something bigger than 1x1 then a
reference to a region will be returned. You can also add horizontal and vertical shifts to the region using the shift
argument.
}
\author{
Ben Veal
}
