% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_helpers.R
\name{isOverlapping}
\alias{isOverlapping}
\title{Check if two regions overlap}
\usage{
isOverlapping(region1, region2)
}
\arguments{
\item{region1}{A vector of row & column indices or an excel reference string for a single cell or a region.}

\item{region2}{A vector of row & column indices or an excel reference string for a region.}
}
\value{
TRUE if region1 overlaps region2
}
\description{
Check if two regions overlap.
}
\details{
Each region argument can be either a numeric vector or matrix of row & column indices (as returned by aref2idx
or getReferenceCoordinatesForName) or an excel reference string (as returned by idx2aref).
Also the first argument (region1) may refer to a single cell instead of a region (i.e. it may be a vector of length 2).
}
\author{
Ben Veal
}
