% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_helpers.R
\name{isWithin}
\alias{isWithin}
\title{Check if a cell/region is within another region}
\usage{
isWithin(cell, region)
}
\arguments{
\item{cell}{The index or aref of a single cell/region}

\item{region}{The index or aref of a region}
}
\value{
TRUE if cell is within region
}
\description{
Check if a cell/region is within another region.
}
\details{
This function returns TRUE if the first arg refers to a cell or region within the region referred to by the
second arg. Each arg can be either a vector of numeric row & column indices, or an excel reference/region string.
}
\author{
Joe Bloggs
}
