% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_template.R
\name{load_xl_template}
\alias{load_xl_template}
\title{Load an Excel reporting template}
\usage{
load_xl_template(template_name)
}
\arguments{
\item{template_name}{path to Excel template}
}
\value{
a list with a reference to the workbook object and vector of
template names.
}
\description{
Returns a list with a reference to the workbook object and vector of
template names.
}
\examples{

\dontrun{
report_obj <- load_xl_template("templates/report_template.xlsx")
}
}
