% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_helpers.R
\name{makeRegionFunction}
\alias{makeRegionFunction}
\title{Create function for writing to worksheet}
\usage{
makeRegionFunction(wb, sheetname, defaultRow = TRUE, defaultCol = TRUE,
  defaultHeader = TRUE)
}
\arguments{
\item{wb}{An excel workbook object}

\item{sheetname}{The name of a worksheet in the workbook object wb}

\item{defaultRow}{Default logical value to use for absRow arg of returned function which indicates whether
rows of cellrefs should be absolute references or not.}

\item{defaultCol}{Default logical value to use for absCol arg of returned function which indicates whether
columns of cellrefs should be absolute references or not.}

\item{defaultHeader}{Default logical value of header arg of returned function which indicates whether column
names of data should be written to the excel file.}
}
\description{
Create a function for writing named regions to a predefined worksheet.
}
\details{
This function returns another convenience function for writing named regions to a given worksheet.
The returned function takes a region name, cell references and a dataframe as its main arguments and writes
the dataframe to the workbook & worksheet pair that were given as arguments to the parent function.
If the worksheet does not yet exist it is created by the parent function. Similarly if the named region does
not yet exist it is created by the child function.
The child function also has named arguments 'header' (see 'writeNamedRegion') and 'absRow' & 'absCol' (see 'idx2cref').
The default values for these arguments can be set using 'defaultHeader', 'defaultRow', and 'defaultCol' arguments of
the parent function.
}
\author{
Ben Veal
}
