% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_helpers.R
\name{rel2absref}
\alias{rel2absref}
\title{Create an absolute reference from base and relative references}
\usage{
rel2absref(baseref, relref)
}
\arguments{
\item{baseref}{An absolute reference to an excel worksheet cell/area in either string form, matrix form or numeric vector form.}

\item{relref}{A relative reference to an excel worksheet cell/area in either string form, matrix form or numeric vector form.}
}
\value{
A vector of 2 or 4 row & column indices for a single cell or an area, with entries 1 & 3 indicating the first & last
rows, and entries 2 & 4 indicating the first and last columns.
}
\description{
Given a cell reference that is relative to a given base reference, create a corresponding absolute cell reference.
}
\details{
The reference arguments can refer to single cells or areas. 
The size of the returned reference area will correspond with the size of the relref argument, and if the relref argument
is a single cell then the returned reference will also be a single cell.
Reference arguments can be either numeric vectors or matrices of row & column indices (as returned by aref2idx
or getReferenceCoordinatesForName) or a excel reference strings (as returned by idx2aref).
A relative reference of 1 corresponds to the first row/column.
Negative indices in the relfref argument count backwards from the bottom/right side of the base reference cell/area,
with 0 corresponding to the last row/column.
}
\examples{
rel2absref("B2:E5","A1:B2")
rel2absref("B2:E5",c(1,1,2,2))
rel2absref("B2:E5",c(1,1,-2,-2))
rel2absref("B2:E5",c(-3,-3,-2,-2))
}
\author{
Ben Veal
}
